+++ src/autometrics/tracker/test_tracker.py
@@ -1,7 +1,10 @@
+from prometheus_client.exposition import generate_latest
+import pytest
+
 from .opentelemetry import OpenTelemetryTracker
 from .prometheus import PrometheusTracker
 
-from .tracker import default_tracker
+from .tracker import default_tracker, init_tracker, TrackerType
 
 
 def test_default_tracker(monkeypatch):
@@ -22,3 +25,55 @@ def test_default_tracker(monkeypatch):
     monkeypatch.setenv("AUTOMETRICS_TRACKER", "something_else")
     tracker = default_tracker()
     assert isinstance(tracker, OpenTelemetryTracker)
+
+
+def test_init_prometheus_tracker_set_build_info(monkeypatch):
+    """Test that init_tracker (for a Prometheus tracker) calls set_build_info using env vars."""
+
+    commit = "d6abce3"
+    version = "1.0.1"
+
+    monkeypatch.setenv("AUTOMETRICS_COMMIT", commit)
+    monkeypatch.setenv("AUTOMETRICS_VERSION", version)
+
+    prom_tracker = init_tracker(TrackerType.PROMETHEUS)
+    assert isinstance(prom_tracker, PrometheusTracker)
+
+    blob = generate_latest()
+    assert blob is not None
+    data = blob.decode("utf-8")
+
+    prom_build_info = f"""build_info{{commit="{commit}",version="{version}"}} 1.0"""
+    assert prom_build_info in data
+
+    monkeypatch.delenv("AUTOMETRICS_VERSION", raising=False)
+    monkeypatch.delenv("AUTOMETRICS_COMMIT", raising=False)
+
+
+def test_init_otel_tracker_set_build_info(monkeypatch):
+    """
+    Test that init_tracker (for an OTEL tracker) calls set_build_info using env vars.
+    Note that the OTEL collector translates metrics to Prometheus.
+    """
+    pytest.skip(
+        "Skipping test because OTEL collector does not create a gauge when it translates UpDownCounter to Prometheus"
+    )
+
+    commit = "a29a178"
+    version = "0.0.1"
+
+    monkeypatch.setenv("AUTOMETRICS_COMMIT", commit)
+    monkeypatch.setenv("AUTOMETRICS_VERSION", version)
+
+    otel_tracker = init_tracker(TrackerType.OPENTELEMETRY)
+    assert isinstance(otel_tracker, OpenTelemetryTracker)
+
+    blob = generate_latest()
+    assert blob is not None
+    data = blob.decode("utf-8")
+
+    prom_build_info = f"""build_info{{commit="{commit}",version="{version}"}} 1.0"""
+    assert prom_build_info in data
+
+    monkeypatch.delenv("AUTOMETRICS_VERSION", raising=False)
+    monkeypatch.delenv("AUTOMETRICS_COMMIT", raising=False)