+++ src/autometrics/tracker/tracker.py
@@ -16,6 +16,9 @@ class Result(Enum):
 class TrackMetrics(Protocol):
     """Protocol for tracking metrics."""
 
+    def set_build_info(self, commit: str, version: str):
+        """Observe the build info. Should only be called once per tracker instance"""
+
     def finish(
         self,
         start_time: float,
@@ -35,18 +38,28 @@ class TrackerType(Enum):
     PROMETHEUS = "prometheus"
 
 
-def create_tracker(tracker_type: TrackerType) -> TrackMetrics:
+def init_tracker(tracker_type: TrackerType) -> TrackMetrics:
     """Create a tracker"""
+
+    tracker_instance: TrackMetrics
     if tracker_type == TrackerType.OPENTELEMETRY:
         # pylint: disable=import-outside-toplevel
         from .opentelemetry import OpenTelemetryTracker
 
-        return OpenTelemetryTracker()
+        tracker_instance = OpenTelemetryTracker()
     elif tracker_type == TrackerType.PROMETHEUS:
         # pylint: disable=import-outside-toplevel
         from .prometheus import PrometheusTracker
 
-        return PrometheusTracker()
+        tracker_instance = PrometheusTracker()
+
+    # NOTE - Only set the build info when the tracker is initialized
+    tracker_instance.set_build_info(
+        commit=os.getenv("AUTOMETRICS_COMMIT") or "",
+        version=os.getenv("AUTOMETRICS_VERSION") or "",
+    )
+
+    return tracker_instance
 
 
 def get_tracker_type() -> TrackerType:
@@ -60,7 +73,7 @@ def get_tracker_type() -> TrackerType:
 def default_tracker():
     """Setup the default tracker."""
     preferred_tracker = get_tracker_type()
-    return create_tracker(preferred_tracker)
+    return init_tracker(preferred_tracker)
 
 
 tracker: TrackMetrics = default_tracker()
@@ -74,4 +87,4 @@ def get_tracker() -> TrackMetrics:
 def set_tracker(tracker_type: TrackerType):
     """Set the tracker type."""
     global tracker
-    tracker = create_tracker(tracker_type)
+    tracker = init_tracker(tracker_type)